move the Exec API in its own package
