+++ ducc/exec/exec.go
@@ -1,4 +1,4 @@
-package lib
+package exec
 
 import (
 	"bytes"
@@ -114,64 +114,6 @@ func (e *execCmd) Start() error {
 	return nil
 }
 
-/*
-func (e *execCmd) Start() error {
-	if e == nil {
-		err := fmt.Errorf("Use of nil execCmd")
-		LogE(err).Error("Call start with nil cmd, maybe error in the constructor")
-		return err
-	}
-
-	err := e.cmd.Start()
-	if err != nil {
-		LogE(err).Error("Error in starting the command")
-		return err
-	}
-
-	if e.stdin != nil {
-		go func() {
-			defer (*e.stdin).Close()
-			defer e.in.Close()
-			n, err := io.Copy(e.in, *e.stdin)
-			Log().WithFields(log.Fields{"n": n}).Info("Copied n bytes to STDIN")
-			if err != nil {
-				LogE(err).Error("Error in copying the input into STDIN.")
-				return
-			}
-			for n > 0 {
-				n, err = io.Copy(e.in, *e.stdin)
-				if err != nil {
-					LogE(err).Error("Error in copying the input into STDIN")
-					return
-				}
-				Log().WithFields(log.Fields{"n": n}).Info("Copied additionally n bytes to STDIN")
-			}
-		}()
-	}
-
-	slurpOut, errOUT := ioutil.ReadAll(e.out)
-	if errOUT != nil {
-		LogE(errOUT).Warning("Impossible to read the STDOUT")
-		return err
-	}
-	slurpErr, errERR := ioutil.ReadAll(e.err)
-	if errERR != nil {
-		LogE(errERR).Warning("Impossible to read the STDERR")
-		return err
-	}
-
-	err = e.cmd.Wait()
-	if err != nil {
-
-		LogE(err).Error("Error in executing the command")
-		Log().WithFields(log.Fields{"pipe": "STDOUT"}).Info(string(slurpOut))
-		Log().WithFields(log.Fields{"pipe": "STDERR"}).Info(string(slurpErr))
-		return err
-	}
-	return nil
-}
-*/
-
 func (e *execCmd) Env(key, value string) *execCmd {
 	if e == nil {
 		err := fmt.Errorf("Use of nil execCmd")
@@ -180,4 +122,12 @@ func (e *execCmd) Env(key, value string) *execCmd {
 	}
 	e.cmd.Env = append(e.cmd.Env, fmt.Sprintf("%s=%s", key, value))
 	return e
+}
+
+func LogE(err error) *log.Entry {
+	return log.WithFields(log.Fields{"error": err})
+}
+
+func Log() *log.Entry {
+	return log.WithFields(log.Fields{})
 }
